wip - start moving swagger core functions to shared
